# onidelcloud/onidel-api-docs.yaml
openapi: 3.0.3
info:
  title: Onidel Cloud API - OpenAPI 3.0
  description: |-
    Welcome to the **Onidel Cloud API** documentation. This API allows developers to seamlessly integrate with the Onidel Cloud platform, enabling them to manage cloud resources efficiently and securely. With our API, you can automate processes, retrieve detailed account information, and manage virtual machines (VPS), networking, storage, and more.
  termsOfService: https://docs.onidel.com/policies/terms-of-service
  contact:
    email: support@onidel.com
  version: 1.2.0
servers:
  - url: https://api.cloud.onidel.com
...
paths:
...
  /network/vpcs:
    get:
      tags:
        - VPC
      summary: List VPCs
      description: Retrieve all VPCs for a team
      operationId: getVPCs
      parameters:
        - name: team_id
          in: query
          description: Team ID to filter VPCs. Default team will be used if not provided.
          required: false
          schema:
            type: string
            format: uuid
        - name: location
          in: query
          description: Filter VPCs by location
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VPCs'
        '401':
          description: Unauthorized
      security:
        - api_key: []
    post:
      tags:
        - VPC
      summary: Create a new VPC
      description: Create a new Virtual Private Cloud network
      operationId: createVPC
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewVPC'
      responses:
        '201':
          description: VPC created successfully
        '400':
          description: Bad request - missing or invalid parameters
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - minimum VM required or max VPC limit reached
        '404':
          description: Location not found
        '503':
          description: Service unavailable - resource not available
      security:
        - api_key: []
...
  /vm:
    get:
      tags:
        - Virtual Machine
      summary: List VMs
      description: List all VMs available to the user
      operationId: getVMs
      parameters:
        - name: team_id
          in: query
          description: Team ID that need to be considered for filter. Default team will be used if not provided.
          required: false
          explode: true
          schema:
            type: string
      security:
        - api_key: []
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VMs'
        '401':
          description: Unauthorized
        '400':
          description: Bad request
    post:
      tags:
        - Virtual Machine
      summary: Provision a new VM
      description: Provision a new VM
      operationId: provisionVM
      security:
        - api_key: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewVM'
      responses:
        '201':
          description: Successful operation
        '400':
          description: Bad request
        '402':
          description: Payment required
        '403':
          description: Forbidden
        '404':
          description: Not found
...
  /startup_scripts:
    get:
      tags:
        - Startup Script
      summary: List Startup Scripts
      description: Retrieve all startup scripts for a team
      operationId: getStartupScripts
      parameters:
        - name: team_id
          in: query
          description: Team ID to filter scripts. Default team will be used if not provided.
          required: false
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StartupScripts'
        '401':
          description: Unauthorized
        '404':
          description: Not found
      security:
        - api_key: []
    post:
      tags:
        - Startup Script
      summary: Create a new Startup Script
      description: Add a new startup script to the team. Maximum 10 scripts per team.
      operationId: createStartupScript
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewStartupScript'
      responses:
        '201':
          description: Startup script created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StartupScriptDetailResponse'
        '400':
          description: Bad request - missing parameters or script limit reached (max 10 per team)
        '401':
          description: Unauthorized
      security:
        - api_key: []
...
components:
  schemas:
...
    VPCTeam:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: c1d45377-b2a9-4407-ab8d-6909c34dfaac
        name:
          type: string
          example: "IT Development"
    VPC:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: f5a3e2d8-9876-4abc-1234-543210fedcba
        name:
          type: string
          example: "Production VPC"
        description:
          type: string
          example: "VPC for production environment"
        date_created:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
        status:
          type: string
          enum:
            - active
            - building
            - deleting
          example: "active"
        location:
          type: string
          example: "sydney"
        v4_subnet:
          type: string
          example: "10.0.0.0"
        v4_subnet_mask:
          type: string
          example: "24"
        owner_team:
          $ref: '#/components/schemas/VPCTeam'
        teams:
          type: array
          items:
            $ref: '#/components/schemas/VPCTeam'
        attached_vms_count:
          type: integer
          example: 5
      required:
        - id
        - name
        - description
        - date_created
        - status
        - location
        - v4_subnet
        - v4_subnet_mask
        - owner_team
        - teams
    VPCs:
      type: object
      properties:
        vpcs:
          type: array
          items:
            $ref: '#/components/schemas/VPC'
...
    NewVPC:
      type: object
      properties:
        team_id:
          type: string
          format: uuid
          example: c1d45377-b2a9-4407-ab8d-6909c34dfaac
          description: "Team ID. Default team will be used if not provided."
        location:
          type: string
          example: "sydney"
        name:
          type: string
          example: "Production VPC"
          maxLength: 200
        description:
          type: string
          example: "VPC for production environment"
          maxLength: 200
        v4_subnet:
          type: string
          example: "10.0.0.0"
          description: "IPv4 subnet prefix"
        v4_subnet_mask:
          type: string
          example: "24"
          description: "IPv4 subnet mask"
      required:
        - location
        - name
        - description
        - v4_subnet
        - v4_subnet_mask
...
    VM:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: c1d45377-b2a9-4407-ab8d-6909c34dfaac
        name:
          type: string
          format: string
          example: "ubuntu-vm-4vcpu-8gb"
        vcpu:
          type: integer
          format: int64
          example: 2
        ram:
          type: integer
          format: int64
          example: 4096
        disk:
          type: integer
          format: int64
          example: 40
        bw_used:
          type: number
          format: float
          example: 12.5
        main_ipv4:
          type: string
          format: string
          example: 160.22.79.122
        main_ipv6:
          type: string
          format: string
          example: "2401:a4a0:1:90::1"
        template:
          type: string
          format: string
          example: "Ubuntu 24.04 x64 LTS"
        firewall_group_id:
          type: string
          format: uuid
          nullable: true
          example: ac135def-c613-4a08-fd8e-a20acbdd23a9
        created_at:
          type: string
          format: datetime
          example: "2024-10-04T10:20:22.922772Z"
        renewed_at:
          type: string
          format: datetime
          example: "2024-10-04T10:20:22.922772Z"
        due_date:
          type: string
          format: datetime
          example: "2024-10-04T10:20:22.922772Z"
        recurring_amount:
          type: number
          format: float
          example: 5.5
        amount_paid:
          type: number
          format: float
          example: 1.2345
        payment_currency:
          type: string
          format: string
          example: "AUD"
        billing_cycle:
          type: integer
          format: int64
          example: 1
        status:
          type: string
          format: string
          enum:
            - building
            - active
            - suspended
            - awaiting_payment
            - taking_snaphot
            - restoring
            - terminating
            - migrating
...
    VMs:
      type: array
      items:
        $ref: '#/components/schemas/VM'
    NewVM:
      type: object
      properties:
        name:
          type: string
          format: string
          example: "vm-hostname"
        payment_cycle:
          type: string
          format: string
          enum:
            - hourly
            - monthly
            - quarterly
            - semiannually
            - annually
            - biennially
            - triennially
        instance_type:
          type: string
          format: uuid
          example: ea645d53-c0e5-406d-8863-9ecd4235be94
        location:
          type: string
          format: string
          example: Sydney
        cpu:
          type: integer
          format: int64
          example: 2
        ram:
          type: integer
          format: int64
          example: 4096
        disk:
          type: integer
          format: int64
          example: 40
        os:
          type: integer
          format: int64
          example: 2
        snapshot_id:
          type: string
          format: uuid
          example: 53959c6e-73d5-4476-9b19-d78fb8efac2c
        team_id:
          type: string
          format: uuid
          example: 6484a164-d355-4947-ac2c-e8c09fb24fb0
        ssh_keys:
          type: array
          items:
            type: string
            format: uuid
          example: ["a3f2e5d8-1234-4abc-9876-543210fedcba"]
          description: "List of SSH key IDs to be added to the VM for authentication"
        vpcs:
          type: array
          items:
            type: string
            format: uuid
          example: ["f5a3e2d8-9876-4abc-1234-543210fedcba"]
          description: "List of VPC IDs to attach the VM to for private networking"
        firewall_group_id:
          type: string
          format: uuid
          example: "e4d3c2b1-5678-9abc-def0-fedcba098765"
          description: "Firewall group ID to apply to the VM for security rules"
        ipv6:
          type: boolean
          example: false
          description: "Enable IPv6 for the VM (if available in the location)"
        disable_ssh_blocking:
          type: boolean
          example: false
          description: "Disable outgoing SSH blocking on the VM (allows outbound SSH connections on port 22)"
        startup_script_id:
          type: string
          format: uuid
          example: "b7e3f1a2-5678-4def-abcd-123456789abc"
          description: "Startup script ID to execute on first boot (only applies when deploying with OS or snapshot)"
...
    StartupScriptSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: b7e3f1a2-5678-4def-abcd-123456789abc
        name:
          type: string
          example: "Install NGINX"
        created:
          type: string
          format: date-time
          example: "2024-06-15T08:00:00Z"
        updated:
          type: string
          format: date-time
          example: "2024-06-15T08:00:00Z"
      required:
        - id
        - name
        - created
        - updated
    StartupScriptDetail:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: b7e3f1a2-5678-4def-abcd-123456789abc
        name:
          type: string
          example: "Install NGINX"
        content:
          type: string
          example: "#!/bin/bash\napt-get update && apt-get install -y nginx"
        created:
          type: string
          format: date-time
          example: "2024-06-15T08:00:00Z"
        updated:
          type: string
          format: date-time
          example: "2024-06-15T08:00:00Z"
      required:
        - id
        - name
        - content
        - created
        - updated
    StartupScripts:
      type: object
      properties:
        scripts:
          type: array
          items:
            $ref: '#/components/schemas/StartupScriptSummary'
    StartupScriptDetailResponse:
      type: object
      properties:
        script:
          $ref: '#/components/schemas/StartupScriptDetail'
    NewStartupScript:
      type: object
      properties:
        team_id:
          type: string
          format: uuid
          example: c1d45377-b2a9-4407-ab8d-6909c34dfaac
        name:
          type: string
          example: "Install NGINX"
        content:
          type: string
          example: "#!/bin/bash\napt-get update && apt-get install -y nginx"
      required:
        - name
        - content
...
  securitySchemes:
    api_key:
      type: http
      scheme: bearer